requirements.txt
Flask==3.0.2
PyYAML==6.0.2
Werkzeug==3.0.1
gunicorn==21.2.0   
pytest==8.0.0      
redis==5.0.1
prometheus_client
python-hcl2